feat: Migration of project template from pranavmishra90/datalad-dataset-template
In order to improve efficiency and maximize reproducibility of research, I have created a template repository with an optimized method for sorting data files (raw data, analyzed data, research output) and the code which analyzes the data. From here, the platform also allows for an electronic lab notebook (ELN) to be created with both static and dynamic components. The environment can be on bare metal or in a dockerized environment using the FacsimiLab platform, pranavmishra90/facsimilab-full.